Extract run function from index.js
